refactor: split widget.py into focused modules (monitors/compositor/desktop/ui)
The single 391-line `widget.py` mixed Qt UI, xrandr parsing, Pillow
compositing, gsettings/dark-mode/dependency-validation, and `main()`
in one class. Split into four small modules with clear seams so the
non-Qt logic is unit-testable without spinning up a QApplication:

* `monitors.py`  - `Monitor` TypedDict, `parse_xrandr_output(stdout)`
                   pure parser, `get_monitors()` subprocess wrapper.
* `compositor.py`- `compute_canvas_size(monitors)` (pure) and
                   `compose_wallpaper(monitors, image_paths, output_paths)`
                   that does the Pillow IO. Output paths are passed in,
                   so writing to multiple locations on GNOME is just a
                   list rather than an inline branch.
* `desktop.py`   - `detect_desktop_environment`, `wallpaper_output_paths`,
                   `is_system_in_dark_mode`, `validate_dependencies`,
                   `apply_background`. All gsettings/XDG concerns live
                   here; no Qt imports.
* `ui.py`        - `MultiMonitorApp(QMainWindow)` and `main()`. UI now
                   *calls into* the three modules above instead of
                   owning that logic as instance methods.

Wiring updates:
* `[project.scripts]` console entry now points at `ui:main`.
* `__init__.py` and `__main__.py` re-export `main` from `.ui`.
* `tests/test_basic.py` imports `MultiMonitorApp` from `.ui`.

Removed (pre-existing dead code, called out for the record):
* `get_screen_geometry()` had no callers anywhere in the codebase.

No behavior change at runtime: `multimonitor-wallpapers`,
`python -m multimonitor_wallpapers`, and `from multimonitor_wallpapers
import main` all still work.
