+++ src/multimonitor_wallpapers/__init__.py
@@ -6,6 +6,6 @@ A Python tool to set different wallpapers on multiple monitors in Linux Cinnamon
 
 __version__ = "0.1.0"
 
-from .widget import main  # noqa
+from .ui import main  # noqa: F401
 
 __all__ = ["main"]
