+++ src/multimonitor_wallpapers/__main__.py
@@ -2,7 +2,7 @@
 Entry point for the MultiMonitor Wallpapers application.
 """
 
-from .widget import main
+from .ui import main
 
 if __name__ == "__main__":
     main()
